requirements.txt
streamlit
python-dotenv

langchain-core
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-groq

sentence-transformers
chromadb
pypdf
Pillow
